Reduce duplication for the Vay pusher on CPU and GPU
